Better handling of the OmniAuth uid value
Previously the OmniAuth uid was provided as the SAML NameID which
is not ideal as with Suomi.fi that is only unique per session.
With this change, the uid value is unique per person, as long as
it can be defined from one of the attributes tied to the
authenticated person.

If it cannot be determined from the person's data, the SAML NameID
is used as a fallback.