Allow application handle the SLO process initiated through another service
In case another service initiates the logout process, the SAML uid
(i.e. the NameID) is not transferred over the current session
because the request happens from an iframe.

Therefore, allow the application itself to handle this situation
by calling the application in this situation. This way, the
application itself can handle the logout of the user using its
internal logic e.g. by storing the sessions in a database.